Refactored the account page into an admin view also
